requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-community
streamlit
unstructured
unstructured[md]
duckduckgo-search
watchdog
pytest
langgraph
strands-agents
strands-agents-tools
ddgs
wikipedia